+++ CHANGELOG.md
@@ -4,3 +4,6 @@ CHANGELOG.md
 + version 1.4.4 added onSuspense and fix out- (document event)
 + version 1.4.7 - 1.4.10 solving resuming issue when wrapping with template
 + version 1.4.11 - added more event modifiers
++ version 1.4.12 - fixed pluginsMap pawaAttribute to pawa-ssr 1.3.3
++ version 1.4.13 - 1.4.16 fixed issues for resuming component 
++ version 1.4.18  fixed issue with component inside template and added event modifiers